# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventarios de activos",
        "description": "APIs para el manejo de inventarios de activos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/digital/api/actions/{id}": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Get an action",
                "description": "Get details of an action from the system",
                "operationId": "GetAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the action",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Actions"
                ],
                "summary": "Delete an action",
                "description": "Delete an action from the system",
                "operationId": "DeleteAction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the action",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/digital/api/criticality-levels": {
            "post": {
                "tags": [
                    "CriticalityLevels"
                ],
                "summary": "Store a new criticality level",
                "description": "Store a new criticality level in the system",
                "operationId": "StoreCriticalityLevel",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "criticality_name"
                                ],
                                "properties": {
                                    "criticality_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/digital/api/criticality-levels/{id}": {
            "get": {
                "tags": [
                    "CriticalityLevels"
                ],
                "summary": "Get a criticality level",
                "description": "Get details of a criticality level from the system",
                "operationId": "GetCriticalityLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the criticality level",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CriticalityLevels"
                ],
                "summary": "Delete a criticality level",
                "description": "Delete a criticality level from the system",
                "operationId": "DeleteCriticalityLevel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the criticality level",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/digital/api/departments": {
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "Store a new department",
                "description": "Store a new department in the system",
                "operationId": "StoreDepartment",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "department_name"
                                ],
                                "properties": {
                                    "department_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/digital/api/departments/{id}": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Get a department",
                "description": "Get details of a department from the system",
                "operationId": "GetDepartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the department",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Departments"
                ],
                "summary": "Delete a department",
                "description": "Delete a department from the system",
                "operationId": "DeleteDepartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the department",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No data found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}